Move Product Collection Inner Blocks structure to constants
